+++ apps/antalmanac/src/components/RightPane/CoursePane/SearchForm/DepartmentSearchBar/constants.ts
